Future Publishing & New Article Options
Articles now:
 - support Future Publishing
 - have a Clear Cache checkbox
 - have a Rich Text checkbox

Previously when a future pub_date was set for an Article, it's resulting URLs wouldn't be generated correctly (ignored Articles URL Format, stays that of the preview URL). Now when an Article is saved, if it has a pub_date it's URI is set accordingly.

Like with Resources, unpicking the Clear Cache checkbox will prevent caches from being cleared on save.

When new Articles are created they still inherit the Rich Text Setting from their Article Container parent. Once created, a Rich Text checkbox will appear in a separate Editing Options panel that allows it to inherit the parent value.

screenshot:
http://j4p.us/image/3i3o2F2v0L0p
Signed-off-by: JP DeVries <johnpaul.devries@gmail.com>